broke the mcp bridge out into multiple files to make mangement better. Cleaned up the code. Added uv support. Changed to dynamic search of tools rather then including all ~200 and poluting the context window.
