+++ .gitignore
@@ -272,6 +272,7 @@ replay_pid*
 *.gpr
 *.rep/
 *.lock
+!uv.lock
 
 # Ghidra analysis files
 *.gzf
