+++ python/ghidra_mcp_bridge/__init__.py
@@ -1,0 +1,5 @@
+"""GhidraMCP bridge — a thin MCP↔HTTP multiplexer for Ghidra instances."""
+
+from .server import main
+
+__all__ = ["main"]
